+++ tests/src/test/scala/slinky/core/ExportedComponentTest.scala
@@ -31,6 +31,11 @@ object TestExportedComponentStateless extends StatelessComponentWrapper {
   }
 }
 
+object TestExportedExternalComponent extends ExternalComponentNoProps {
+  case class Props(children: Seq[ReactElement])
+  val component = "div"
+}
+
 class ExportedComponentTest extends AnyFunSuite {
   test("Can construct an instance of an exported component with JS-provided props") {
     val container = document.createElement("div")
@@ -71,5 +76,17 @@ class ExportedComponentTest extends AnyFunSuite {
     ), container)
 
     assert(container.innerHTML == "lol")
+  }
+
+  test("Can construct an instance of an exported external component with JS-provided props") {
+    val container = document.createElement("div")
+    ReactDOM.render(React.createElement(
+      TestExportedExternalComponent: ReactComponentClass[_],
+      js.Dictionary(
+        "children" -> js.Array("hello")
+      )
+    ), container)
+
+    assert(container.innerHTML == "<div>hello</div>")
   }
 }